Implement winning & losing conditions
Compare the character the user inputted with the characters of the word
to find. If there is a match, the matching characters are revealed. If
there is no matching character, the player lose one life. If all word
characters are found, it's a win. If the player lose all his lives, it's
a loss. In both cases, the game stops at this point.