Allow the player to restart a game + fix
The player can now restart a game by pressing Y after the game is
finished. Fix a floating point (core dump) exception that was caused by
the dictionary file that was not closed before trying to open it again.
Initialize more variables at the start of the main script for a cleaner
code.